Renames
1. `ops` -> `succ-op`

`ops` is used to represent an operation which is evaluated in upper stage on
building; it means `ops` will be funcalled later than present building
stage on execution. I think `succ-op` fit more than `ops`.

2. `next-op` -> `op`

`next-op` is a operation which will be funcalled at execution on the
present stage. It's not 'next' operation.